rename integrations_app for accommodating flagships
