+++ .azure/app-cloud-e2e.yml
@@ -31,7 +31,7 @@ pr:
       - ".azure/app-cloud-e2e.yml"
       - "src/lightning_app/**"
       - "requirements/app/**"
-      - "tests/integrations_app_examples/**"
+      - "tests/integrations_app/**"
       - "setup.py"
     exclude:
       - "requirements/*/docs.txt"
@@ -171,7 +171,7 @@ jobs:
         python -m pytest ${TEST_FILE}::test_${TEST_APP_NAME}_example_cloud \
           --timeout=540 --capture=no -v --color=yes
       env:
-        TEST_FILE: tests/integrations_app_examples/$(TEST_APP_FOLDER)/test_$(TEST_APP_NAME).py
+        TEST_FILE: tests/integrations_app/$(TEST_APP_FOLDER)/test_$(TEST_APP_NAME).py
         #LAI_USER: $(LAI_USER)  # for STAGING
         #LAI_PASS: $(LAI_PASS)  # for STAGING
         LIGHTNING_USER_ID: $(LIGHTNING_USER_ID_PROD)
