+++ .github/checkgroup.yml
@@ -245,7 +245,7 @@ subprojects:
       - ".actions/**"
       - ".github/workflows/ci-examples-app.yml"
       - "src/lightning_app/**"
-      - "tests/integrations_app_examples/**"
+      - "tests/integrations_app/**"
       - "examples/app_*/**"
       - "requirements/app/**"
       - "setup.py"
@@ -269,7 +269,7 @@ subprojects:
       - ".azure/app-cloud-e2e.yml"
       - "src/lightning_app/**"
       - "requirements/app/**"
-      - "tests/integrations_app_examples/**"
+      - "tests/integrations_app/**"
       - "setup.py"
       - "!requirements/*/docs.txt"
       - "!*.md"
