+++ tests/integrations_app/apps/collect_failures/requirements.txt
