+++ tests/integrations_app/local/test_core_features_app.py
@@ -1,24 +1,27 @@
 import os
 
 from click.testing import CliRunner
-from integrations_app_examples.public import _PATH_EXAMPLES
+from integrations_app.local import _PATH_APPS
 
 from lightning_app.cli.lightning_cli import run_app
 
 
-def test_layout_example():
+def test_core_features_app_example():
 
     runner = CliRunner()
     result = runner.invoke(
         run_app,
         [
-            os.path.join(_PATH_EXAMPLES, "app_layout", "app.py"),
+            os.path.join(_PATH_APPS, "core_features_app", "app.py"),
             "--blocking",
             "False",
             "--open-ui",
             "False",
+            "--env",  # this is to test env variable
+            "FOO=bar",
+            "--env",
+            "BLA=bloz",
         ],
         catch_exceptions=False,
     )
-    assert "Layout End" in str(result.stdout_bytes)
     assert result.exit_code == 0
