+++ tests/integrations_app/local/test_idle_timeout.py
@@ -2,7 +2,7 @@ import os
 from time import sleep
 
 import pytest
-from integrations_app_examples.local import _PATH_APPS
+from integrations_app.local import _PATH_APPS
 
 from lightning_app.testing.testing import run_app_in_cloud
