+++ tests/integrations_app/public/test_installation_commands_app.py
@@ -1,7 +1,7 @@
 import os
 
 import pytest
-from integrations_app_examples.public import _PATH_EXAMPLES
+from integrations_app.public import _PATH_EXAMPLES
 
 from lightning_app.testing.testing import run_app_in_cloud
