+++ tests/integrations_app/public/test_layout.py
@@ -1,27 +1,24 @@
 import os
 
 from click.testing import CliRunner
-from integrations_app_examples.local import _PATH_APPS
+from integrations_app.public import _PATH_EXAMPLES
 
 from lightning_app.cli.lightning_cli import run_app
 
 
-def test_core_features_app_example():
+def test_layout_example():
 
     runner = CliRunner()
     result = runner.invoke(
         run_app,
         [
-            os.path.join(_PATH_APPS, "core_features_app", "app.py"),
+            os.path.join(_PATH_EXAMPLES, "app_layout", "app.py"),
             "--blocking",
             "False",
             "--open-ui",
             "False",
-            "--env",  # this is to test env variable
-            "FOO=bar",
-            "--env",
-            "BLA=bloz",
         ],
         catch_exceptions=False,
     )
+    assert "Layout End" in str(result.stdout_bytes)
     assert result.exit_code == 0
