+++ tests/integrations_app/public/test_quick_start.py
@@ -4,7 +4,7 @@ from unittest import mock
 
 import pytest
 from click.testing import CliRunner
-from integrations_app_examples.public import _PATH_EXAMPLES
+from integrations_app.public import _PATH_EXAMPLES
 
 from lightning_app import LightningApp
 from lightning_app.cli.lightning_cli import run_app
